Add uncommitted-block-ids and ability to delete utxos.
Because of the confusion created by the persistent identifiers
of uncommitted blocks using the same type and space as the
blockchain block identifiers, uncommitted blocks now use a
64-bit identifier that is returned by the call to `open_block`.

The rules for validating the `txo_t` passed to `write_txos` have
been made explicit. In particular, it is possible to *delete*
spending or production events that happened within the current
block.